auth/auth_api/api/v1/schemas/totp_request.py
from auth_api.extensions import ma
from marshmallow.validate import Length


class TOTPRequestSchema(ma.Schema):
    totp_code = ma.String(
        required=True, validate=Length(min=6, max=6), example="411526"
    )
    totp_status = ma.Boolean(required=True)
