auth/requirements-tests.txt
-r requirements.txt
pytest
pytest-runner
pytest-flask
pytest-factoryboy
factory_boy
